+++ src/gpio.rs
@@ -14,6 +14,14 @@ pub trait GpioExt {
     fn split(self) -> Self::Parts;
 }
 
+pub trait PinExt {
+    type Mode;
+    /// Return pin number
+    fn pin_id(&self) -> u8;
+    /// Return port number
+    fn port_id(&self) -> u8;
+}
+
 pub struct AF0;
 pub struct AF1;
 pub struct AF2;
@@ -84,6 +92,15 @@ pub enum Edge {
     RISING_FALLING,
 }
 
+mod sealed {
+    /// Marker trait that show if `ExtiPin` can be implemented
+    pub trait Interruptable {}
+}
+
+use sealed::Interruptable;
+impl<MODE> Interruptable for Output<MODE> {}
+impl<MODE> Interruptable for Input<MODE> {}
+
 /// External Interrupt Pin
 pub trait ExtiPin {
     fn make_interrupt_source(&mut self, syscfg: &mut SysCfg);
@@ -94,151 +111,98 @@ pub trait ExtiPin {
     fn check_interrupt(&self) -> bool;
 }
 
-macro_rules! exti_erased {
-    ($PIN:ty, $extigpionr:expr) => {
-        impl<MODE> ExtiPin for $PIN {
-            /// Make corresponding EXTI line sensitive to this pin
-            fn make_interrupt_source(&mut self, syscfg: &mut SysCfg) {
-                let offset = 4 * (self.i % 4);
-                match self.i {
-                    0..=3 => {
-                        syscfg.exticr1.modify(|r, w| unsafe {
-                            w.bits((r.bits() & !(0xf << offset)) | ($extigpionr << offset))
-                        });
-                    }
-                    4..=7 => {
-                        syscfg.exticr2.modify(|r, w| unsafe {
-                            w.bits((r.bits() & !(0xf << offset)) | ($extigpionr << offset))
-                        });
-                    }
-                    8..=11 => {
-                        syscfg.exticr3.modify(|r, w| unsafe {
-                            w.bits((r.bits() & !(0xf << offset)) | ($extigpionr << offset))
-                        });
-                    }
-                    12..=15 => {
-                        syscfg.exticr4.modify(|r, w| unsafe {
-                            w.bits((r.bits() & !(0xf << offset)) | ($extigpionr << offset))
-                        });
-                    }
-                    _ => {}
-                }
-            }
-
-            /// Generate interrupt on rising edge, falling edge or both
-            fn trigger_on_edge(&mut self, exti: &mut EXTI, edge: Edge) {
-                match edge {
-                    Edge::RISING => {
-                        exti.rtsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << self.i)) });
-                        exti.ftsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << self.i)) });
-                    }
-                    Edge::FALLING => {
-                        exti.ftsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << self.i)) });
-                        exti.rtsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << self.i)) });
-                    }
-                    Edge::RISING_FALLING => {
-                        exti.rtsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << self.i)) });
-                        exti.ftsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << self.i)) });
-                    }
-                }
-            }
-
-            /// Enable external interrupts from this pin.
-            fn enable_interrupt(&mut self, exti: &mut EXTI) {
-                exti.imr
-                    .modify(|r, w| unsafe { w.bits(r.bits() | (1 << self.i)) });
-            }
-
-            /// Disable external interrupts from this pin
-            fn disable_interrupt(&mut self, exti: &mut EXTI) {
-                exti.imr
-                    .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << self.i)) });
-            }
-
-            /// Clear the interrupt pending bit for this pin
-            fn clear_interrupt_pending_bit(&mut self) {
-                unsafe { (*EXTI::ptr()).pr.write(|w| w.bits(1 << self.i)) };
-            }
-
-            /// Reads the interrupt pending bit for this pin
-            fn check_interrupt(&self) -> bool {
-                unsafe { ((*EXTI::ptr()).pr.read().bits() & (1 << self.i)) != 0 }
-            }
-        }
-    };
-}
-
-macro_rules! exti {
-    ($PIN:ty, $extigpionr:expr, $i:expr, $exticri:ident) => {
-        impl<MODE> ExtiPin for $PIN {
-            /// Configure EXTI Line $i to trigger from this pin.
-            fn make_interrupt_source(&mut self, syscfg: &mut SysCfg) {
-                let offset = 4 * ($i % 4);
-                syscfg.$exticri.modify(|r, w| unsafe {
-                    let mut exticr = r.bits();
-                    exticr = (exticr & !(0xf << offset)) | ($extigpionr << offset);
-                    w.bits(exticr)
+impl<PIN> ExtiPin for PIN
+where
+    PIN: PinExt,
+    PIN::Mode: Interruptable,
+{
+    /// Make corresponding EXTI line sensitive to this pin
+    #[inline(always)]
+    fn make_interrupt_source(&mut self, syscfg: &mut SysCfg) {
+        let i = self.pin_id();
+        let port = self.port_id() as u32;
+        let offset = 4 * (i % 4);
+        match i {
+            0..=3 => {
+                syscfg.exticr1.modify(|r, w| unsafe {
+                    w.bits((r.bits() & !(0xf << offset)) | (port << offset))
                 });
             }
-
-            /// Generate interrupt on rising edge, falling edge or both
-            fn trigger_on_edge(&mut self, exti: &mut EXTI, edge: Edge) {
-                match edge {
-                    Edge::RISING => {
-                        exti.rtsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << $i)) });
-                        exti.ftsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << $i)) });
-                    }
-                    Edge::FALLING => {
-                        exti.ftsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << $i)) });
-                        exti.rtsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << $i)) });
-                    }
-                    Edge::RISING_FALLING => {
-                        exti.rtsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << $i)) });
-                        exti.ftsr
-                            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << $i)) });
-                    }
-                }
+            4..=7 => {
+                syscfg.exticr2.modify(|r, w| unsafe {
+                    w.bits((r.bits() & !(0xf << offset)) | (port << offset))
+                });
             }
-
-            /// Enable external interrupts from this pin.
-            fn enable_interrupt(&mut self, exti: &mut EXTI) {
-                exti.imr
-                    .modify(|r, w| unsafe { w.bits(r.bits() | (1 << $i)) });
+            8..=11 => {
+                syscfg.exticr3.modify(|r, w| unsafe {
+                    w.bits((r.bits() & !(0xf << offset)) | (port << offset))
+                });
             }
-
-            /// Disable external interrupts from this pin
-            fn disable_interrupt(&mut self, exti: &mut EXTI) {
-                exti.imr
-                    .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << $i)) });
+            12..=15 => {
+                syscfg.exticr4.modify(|r, w| unsafe {
+                    w.bits((r.bits() & !(0xf << offset)) | (port << offset))
+                });
             }
+            _ => unreachable!(),
+        }
+    }
 
-            /// Clear the interrupt pending bit for this pin
-            fn clear_interrupt_pending_bit(&mut self) {
-                unsafe { (*EXTI::ptr()).pr.write(|w| w.bits(1 << $i)) };
+    /// Generate interrupt on rising edge, falling edge or both
+    #[inline(always)]
+    fn trigger_on_edge(&mut self, exti: &mut EXTI, edge: Edge) {
+        let i = self.pin_id();
+        match edge {
+            Edge::RISING => {
+                exti.rtsr
+                    .modify(|r, w| unsafe { w.bits(r.bits() | (1 << i)) });
+                exti.ftsr
+                    .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << i)) });
             }
-
-            /// Reads the interrupt pending bit for this pin
-            fn check_interrupt(&self) -> bool {
-                unsafe { ((*EXTI::ptr()).pr.read().bits() & (1 << $i)) != 0 }
+            Edge::FALLING => {
+                exti.ftsr
+                    .modify(|r, w| unsafe { w.bits(r.bits() | (1 << i)) });
+                exti.rtsr
+                    .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << i)) });
+            }
+            Edge::RISING_FALLING => {
+                exti.rtsr
+                    .modify(|r, w| unsafe { w.bits(r.bits() | (1 << i)) });
+                exti.ftsr
+                    .modify(|r, w| unsafe { w.bits(r.bits() | (1 << i)) });
             }
         }
-    };
+    }
+
+    /// Enable external interrupts from this pin.
+    #[inline(always)]
+    fn enable_interrupt(&mut self, exti: &mut EXTI) {
+        exti.imr
+            .modify(|r, w| unsafe { w.bits(r.bits() | (1 << self.pin_id())) });
+    }
+
+    /// Disable external interrupts from this pin
+    #[inline(always)]
+    fn disable_interrupt(&mut self, exti: &mut EXTI) {
+        exti.imr
+            .modify(|r, w| unsafe { w.bits(r.bits() & !(1 << self.pin_id())) });
+    }
+
+    /// Clear the interrupt pending bit for this pin
+    #[inline(always)]
+    fn clear_interrupt_pending_bit(&mut self) {
+        unsafe { (*EXTI::ptr()).pr.write(|w| w.bits(1 << self.pin_id())) };
+    }
+
+    /// Reads the interrupt pending bit for this pin
+    #[inline(always)]
+    fn check_interrupt(&self) -> bool {
+        unsafe { ((*EXTI::ptr()).pr.read().bits() & (1 << self.pin_id())) != 0 }
+    }
 }
 
 macro_rules! gpio {
-    ($GPIOX:ident, $gpiox:ident, $PXx:ident, $extigpionr:expr, [
-        $($PXi:ident: ($pxi:ident, $i:expr, $MODE:ty, $exticri:ident),)+
+    ($GPIOX:ident, $gpiox:ident, $PXx:ident, $port_id:expr, [
+        $($PXi:ident: ($pxi:ident, $i:expr, $MODE:ty),)+
     ]) => {
         /// GPIO
         pub mod $gpiox {
@@ -246,14 +210,12 @@ macro_rules! gpio {
             use core::convert::Infallible;
 
             use embedded_hal::digital::v2::{InputPin, OutputPin, StatefulOutputPin, toggleable};
-            use crate::pac::$GPIOX;
-
-            use crate::{pac::{RCC, EXTI}, syscfg::SysCfg};
+            use crate::pac::{$GPIOX, RCC};
             use crate::rcc::Enable;
             use super::{
                 Alternate, AlternateOD, Floating, GpioExt, Input, OpenDrain, Output, Speed,
                 PullDown, PullUp, PushPull, AF0, AF1, AF2, AF3, AF4, AF5, AF6, AF7, AF8, AF9, AF10,
-                AF11, AF12, AF13, AF14, AF15, Analog, Edge, ExtiPin,
+                AF11, AF12, AF13, AF14, AF15, Analog, PinExt,
             };
 
             /// GPIO parts
@@ -289,9 +251,16 @@ macro_rules! gpio {
                 _mode: PhantomData<MODE>,
             }
 
-            impl<MODE> $PXx<MODE> {
-                pub fn get_id(&self) -> u8 {
+            impl<MODE> PinExt for $PXx<MODE> {
+                type Mode = MODE;
+
+                #[inline(always)]
+                fn pin_id(&self) -> u8 {
                     self.i
+                }
+                #[inline(always)]
+                fn port_id(&self) -> u8 {
+                    $port_id
                 }
             }
 
@@ -350,10 +319,6 @@ macro_rules! gpio {
                 }
             }
 
-            exti_erased!($PXx<Output<MODE>>, $extigpionr);
-
-            exti_erased!($PXx<Input<MODE>>, $extigpionr);
-
             fn _set_alternate_mode (index: usize, mode: u32)
             {
                 let offset = 2 * index;
@@ -379,6 +344,19 @@ macro_rules! gpio {
                 /// Pin
                 pub struct $PXi<MODE> {
                     _mode: PhantomData<MODE>,
+                }
+
+                impl<MODE> PinExt for $PXi<MODE> {
+                    type Mode = MODE;
+
+                    #[inline(always)]
+                    fn pin_id(&self) -> u8 {
+                        $i
+                    }
+                    #[inline(always)]
+                    fn port_id(&self) -> u8 {
+                        $port_id
+                    }
                 }
 
                 impl<MODE> $PXi<MODE> {
@@ -823,151 +801,146 @@ macro_rules! gpio {
                         Ok(unsafe { (*$GPIOX::ptr()).idr.read().bits() & (1 << $i) == 0 })
                     }
                 }
-
-                exti!($PXi<Output<MODE>>, $extigpionr, $i, $exticri);
-
-                exti!($PXi<Input<MODE>>, $extigpionr, $i, $exticri);
-
             )+
         }
     }
 }
 
 gpio!(GPIOA, gpioa, PA, 0, [
-    PA0: (pa0, 0, Input<Floating>, exticr1),
-    PA1: (pa1, 1, Input<Floating>, exticr1),
-    PA2: (pa2, 2, Input<Floating>, exticr1),
-    PA3: (pa3, 3, Input<Floating>, exticr1),
-    PA4: (pa4, 4, Input<Floating>, exticr2),
-    PA5: (pa5, 5, Input<Floating>, exticr2),
-    PA6: (pa6, 6, Input<Floating>, exticr2),
-    PA7: (pa7, 7, Input<Floating>, exticr2),
-    PA8: (pa8, 8, Input<Floating>, exticr3),
-    PA9: (pa9, 9, Input<Floating>, exticr3),
-    PA10: (pa10, 10, Input<Floating>, exticr3),
-    PA11: (pa11, 11, Input<Floating>, exticr3),
-    PA12: (pa12, 12, Input<Floating>, exticr4),
-    PA13: (pa13, 13, Input<Floating>, exticr4),
-    PA14: (pa14, 14, Input<Floating>, exticr4),
-    PA15: (pa15, 15, Input<Floating>, exticr4),
+    PA0: (pa0, 0, Input<Floating>),
+    PA1: (pa1, 1, Input<Floating>),
+    PA2: (pa2, 2, Input<Floating>),
+    PA3: (pa3, 3, Input<Floating>),
+    PA4: (pa4, 4, Input<Floating>),
+    PA5: (pa5, 5, Input<Floating>),
+    PA6: (pa6, 6, Input<Floating>),
+    PA7: (pa7, 7, Input<Floating>),
+    PA8: (pa8, 8, Input<Floating>),
+    PA9: (pa9, 9, Input<Floating>),
+    PA10: (pa10, 10, Input<Floating>),
+    PA11: (pa11, 11, Input<Floating>),
+    PA12: (pa12, 12, Input<Floating>),
+    PA13: (pa13, 13, Input<Floating>),
+    PA14: (pa14, 14, Input<Floating>),
+    PA15: (pa15, 15, Input<Floating>),
 ]);
 
 gpio!(GPIOB, gpiob, PB, 1, [
-    PB0: (pb0, 0, Input<Floating>, exticr1),
-    PB1: (pb1, 1, Input<Floating>, exticr1),
-    PB2: (pb2, 2, Input<Floating>, exticr1),
-    PB3: (pb3, 3, Input<Floating>, exticr1),
-    PB4: (pb4, 4, Input<Floating>, exticr2),
-    PB5: (pb5, 5, Input<Floating>, exticr2),
-    PB6: (pb6, 6, Input<Floating>, exticr2),
-    PB7: (pb7, 7, Input<Floating>, exticr2),
-    PB8: (pb8, 8, Input<Floating>, exticr3),
-    PB9: (pb9, 9, Input<Floating>, exticr3),
-    PB10: (pb10, 10, Input<Floating>, exticr3),
-    PB11: (pb11, 11, Input<Floating>, exticr3),
-    PB12: (pb12, 12, Input<Floating>, exticr4),
-    PB13: (pb13, 13, Input<Floating>, exticr4),
-    PB14: (pb14, 14, Input<Floating>, exticr4),
-    PB15: (pb15, 15, Input<Floating>, exticr4),
+    PB0: (pb0, 0, Input<Floating>),
+    PB1: (pb1, 1, Input<Floating>),
+    PB2: (pb2, 2, Input<Floating>),
+    PB3: (pb3, 3, Input<Floating>),
+    PB4: (pb4, 4, Input<Floating>),
+    PB5: (pb5, 5, Input<Floating>),
+    PB6: (pb6, 6, Input<Floating>),
+    PB7: (pb7, 7, Input<Floating>),
+    PB8: (pb8, 8, Input<Floating>),
+    PB9: (pb9, 9, Input<Floating>),
+    PB10: (pb10, 10, Input<Floating>),
+    PB11: (pb11, 11, Input<Floating>),
+    PB12: (pb12, 12, Input<Floating>),
+    PB13: (pb13, 13, Input<Floating>),
+    PB14: (pb14, 14, Input<Floating>),
+    PB15: (pb15, 15, Input<Floating>),
 ]);
 
 gpio!(GPIOC, gpioc, PC, 2, [
-    PC0: (pc0, 0, Input<Floating>, exticr1),
-    PC1: (pc1, 1, Input<Floating>, exticr1),
-    PC2: (pc2, 2, Input<Floating>, exticr1),
-    PC3: (pc3, 3, Input<Floating>, exticr1),
-    PC4: (pc4, 4, Input<Floating>, exticr2),
-    PC5: (pc5, 5, Input<Floating>, exticr2),
-    PC6: (pc6, 6, Input<Floating>, exticr2),
-    PC7: (pc7, 7, Input<Floating>, exticr2),
-    PC8: (pc8, 8, Input<Floating>, exticr3),
-    PC9: (pc9, 9, Input<Floating>, exticr3),
-    PC10: (pc10, 10, Input<Floating>, exticr3),
-    PC11: (pc11, 11, Input<Floating>, exticr3),
-    PC12: (pc12, 12, Input<Floating>, exticr4),
-    PC13: (pc13, 13, Input<Floating>, exticr4),
-    PC14: (pc14, 14, Input<Floating>, exticr4),
-    PC15: (pc15, 15, Input<Floating>, exticr4),
+    PC0: (pc0, 0, Input<Floating>),
+    PC1: (pc1, 1, Input<Floating>),
+    PC2: (pc2, 2, Input<Floating>),
+    PC3: (pc3, 3, Input<Floating>),
+    PC4: (pc4, 4, Input<Floating>),
+    PC5: (pc5, 5, Input<Floating>),
+    PC6: (pc6, 6, Input<Floating>),
+    PC7: (pc7, 7, Input<Floating>),
+    PC8: (pc8, 8, Input<Floating>),
+    PC9: (pc9, 9, Input<Floating>),
+    PC10: (pc10, 10, Input<Floating>),
+    PC11: (pc11, 11, Input<Floating>),
+    PC12: (pc12, 12, Input<Floating>),
+    PC13: (pc13, 13, Input<Floating>),
+    PC14: (pc14, 14, Input<Floating>),
+    PC15: (pc15, 15, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpiod")]
 gpio!(GPIOD, gpiod, PD, 3, [
-    PD0: (pd0, 0, Input<Floating>, exticr1),
-    PD1: (pd1, 1, Input<Floating>, exticr1),
-    PD2: (pd2, 2, Input<Floating>, exticr1),
-    PD3: (pd3, 3, Input<Floating>, exticr1),
-    PD4: (pd4, 4, Input<Floating>, exticr2),
-    PD5: (pd5, 5, Input<Floating>, exticr2),
-    PD6: (pd6, 6, Input<Floating>, exticr2),
-    PD7: (pd7, 7, Input<Floating>, exticr2),
-    PD8: (pd8, 8, Input<Floating>, exticr3),
-    PD9: (pd9, 9, Input<Floating>, exticr3),
-    PD10: (pd10, 10, Input<Floating>, exticr3),
-    PD11: (pd11, 11, Input<Floating>, exticr3),
-    PD12: (pd12, 12, Input<Floating>, exticr4),
-    PD13: (pd13, 13, Input<Floating>, exticr4),
-    PD14: (pd14, 14, Input<Floating>, exticr4),
-    PD15: (pd15, 15, Input<Floating>, exticr4),
+    PD0: (pd0, 0, Input<Floating>),
+    PD1: (pd1, 1, Input<Floating>),
+    PD2: (pd2, 2, Input<Floating>),
+    PD3: (pd3, 3, Input<Floating>),
+    PD4: (pd4, 4, Input<Floating>),
+    PD5: (pd5, 5, Input<Floating>),
+    PD6: (pd6, 6, Input<Floating>),
+    PD7: (pd7, 7, Input<Floating>),
+    PD8: (pd8, 8, Input<Floating>),
+    PD9: (pd9, 9, Input<Floating>),
+    PD10: (pd10, 10, Input<Floating>),
+    PD11: (pd11, 11, Input<Floating>),
+    PD12: (pd12, 12, Input<Floating>),
+    PD13: (pd13, 13, Input<Floating>),
+    PD14: (pd14, 14, Input<Floating>),
+    PD15: (pd15, 15, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpioe")]
 gpio!(GPIOE, gpioe, PE, 4, [
-    PE0: (pe0, 0, Input<Floating>, exticr1),
-    PE1: (pe1, 1, Input<Floating>, exticr1),
-    PE2: (pe2, 2, Input<Floating>, exticr1),
-    PE3: (pe3, 3, Input<Floating>, exticr1),
-    PE4: (pe4, 4, Input<Floating>, exticr2),
-    PE5: (pe5, 5, Input<Floating>, exticr2),
-    PE6: (pe6, 6, Input<Floating>, exticr2),
-    PE7: (pe7, 7, Input<Floating>, exticr2),
-    PE8: (pe8, 8, Input<Floating>, exticr3),
-    PE9: (pe9, 9, Input<Floating>, exticr3),
-    PE10: (pe10, 10, Input<Floating>, exticr3),
-    PE11: (pe11, 11, Input<Floating>, exticr3),
-    PE12: (pe12, 12, Input<Floating>, exticr4),
-    PE13: (pe13, 13, Input<Floating>, exticr4),
-    PE14: (pe14, 14, Input<Floating>, exticr4),
-    PE15: (pe15, 15, Input<Floating>, exticr4),
+    PE0: (pe0, 0, Input<Floating>),
+    PE1: (pe1, 1, Input<Floating>),
+    PE2: (pe2, 2, Input<Floating>),
+    PE3: (pe3, 3, Input<Floating>),
+    PE4: (pe4, 4, Input<Floating>),
+    PE5: (pe5, 5, Input<Floating>),
+    PE6: (pe6, 6, Input<Floating>),
+    PE7: (pe7, 7, Input<Floating>),
+    PE8: (pe8, 8, Input<Floating>),
+    PE9: (pe9, 9, Input<Floating>),
+    PE10: (pe10, 10, Input<Floating>),
+    PE11: (pe11, 11, Input<Floating>),
+    PE12: (pe12, 12, Input<Floating>),
+    PE13: (pe13, 13, Input<Floating>),
+    PE14: (pe14, 14, Input<Floating>),
+    PE15: (pe15, 15, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpiof")]
 gpio!(GPIOF, gpiof, PF, 5, [
-    PF0: (pf0, 0, Input<Floating>, exticr1),
-    PF1: (pf1, 1, Input<Floating>, exticr1),
-    PF2: (pf2, 2, Input<Floating>, exticr1),
-    PF3: (pf3, 3, Input<Floating>, exticr1),
-    PF4: (pf4, 4, Input<Floating>, exticr2),
-    PF5: (pf5, 5, Input<Floating>, exticr2),
-    PF6: (pf6, 6, Input<Floating>, exticr2),
-    PF7: (pf7, 7, Input<Floating>, exticr2),
-    PF8: (pf8, 8, Input<Floating>, exticr3),
-    PF9: (pf9, 9, Input<Floating>, exticr3),
-    PF10: (pf10, 10, Input<Floating>, exticr3),
-    PF11: (pf11, 11, Input<Floating>, exticr3),
-    PF12: (pf12, 12, Input<Floating>, exticr4),
-    PF13: (pf13, 13, Input<Floating>, exticr4),
-    PF14: (pf14, 14, Input<Floating>, exticr4),
-    PF15: (pf15, 15, Input<Floating>, exticr4),
+    PF0: (pf0, 0, Input<Floating>),
+    PF1: (pf1, 1, Input<Floating>),
+    PF2: (pf2, 2, Input<Floating>),
+    PF3: (pf3, 3, Input<Floating>),
+    PF4: (pf4, 4, Input<Floating>),
+    PF5: (pf5, 5, Input<Floating>),
+    PF6: (pf6, 6, Input<Floating>),
+    PF7: (pf7, 7, Input<Floating>),
+    PF8: (pf8, 8, Input<Floating>),
+    PF9: (pf9, 9, Input<Floating>),
+    PF10: (pf10, 10, Input<Floating>),
+    PF11: (pf11, 11, Input<Floating>),
+    PF12: (pf12, 12, Input<Floating>),
+    PF13: (pf13, 13, Input<Floating>),
+    PF14: (pf14, 14, Input<Floating>),
+    PF15: (pf15, 15, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpiog")]
 gpio!(GPIOG, gpiog, PG, 6, [
-    PG0: (pg0, 0, Input<Floating>, exticr1),
-    PG1: (pg1, 1, Input<Floating>, exticr1),
-    PG2: (pg2, 2, Input<Floating>, exticr1),
-    PG3: (pg3, 3, Input<Floating>, exticr1),
-    PG4: (pg4, 4, Input<Floating>, exticr2),
-    PG5: (pg5, 5, Input<Floating>, exticr2),
-    PG6: (pg6, 6, Input<Floating>, exticr2),
-    PG7: (pg7, 7, Input<Floating>, exticr2),
-    PG8: (pg8, 8, Input<Floating>, exticr3),
-    PG9: (pg9, 9, Input<Floating>, exticr3),
-    PG10: (pg10, 10, Input<Floating>, exticr3),
-    PG11: (pg11, 11, Input<Floating>, exticr3),
-    PG12: (pg12, 12, Input<Floating>, exticr4),
-    PG13: (pg13, 13, Input<Floating>, exticr4),
-    PG14: (pg14, 14, Input<Floating>, exticr4),
-    PG15: (pg15, 15, Input<Floating>, exticr4),
+    PG0: (pg0, 0, Input<Floating>),
+    PG1: (pg1, 1, Input<Floating>),
+    PG2: (pg2, 2, Input<Floating>),
+    PG3: (pg3, 3, Input<Floating>),
+    PG4: (pg4, 4, Input<Floating>),
+    PG5: (pg5, 5, Input<Floating>),
+    PG6: (pg6, 6, Input<Floating>),
+    PG7: (pg7, 7, Input<Floating>),
+    PG8: (pg8, 8, Input<Floating>),
+    PG9: (pg9, 9, Input<Floating>),
+    PG10: (pg10, 10, Input<Floating>),
+    PG11: (pg11, 11, Input<Floating>),
+    PG12: (pg12, 12, Input<Floating>),
+    PG13: (pg13, 13, Input<Floating>),
+    PG14: (pg14, 14, Input<Floating>),
+    PG15: (pg15, 15, Input<Floating>),
 ]);
 
 #[cfg(any(
@@ -989,78 +962,78 @@ gpio!(GPIOG, gpiog, PG, 6, [
     feature = "stm32f479"
 ))]
 gpio!(GPIOH, gpioh, PH, 7, [
-    PH0: (ph0, 0, Input<Floating>, exticr1),
-    PH1: (ph1, 1, Input<Floating>, exticr1),
-    PH2: (ph2, 2, Input<Floating>, exticr1),
-    PH3: (ph3, 3, Input<Floating>, exticr1),
-    PH4: (ph4, 4, Input<Floating>, exticr2),
-    PH5: (ph5, 5, Input<Floating>, exticr2),
-    PH6: (ph6, 6, Input<Floating>, exticr2),
-    PH7: (ph7, 7, Input<Floating>, exticr2),
-    PH8: (ph8, 8, Input<Floating>, exticr3),
-    PH9: (ph9, 9, Input<Floating>, exticr3),
-    PH10: (ph10, 10, Input<Floating>, exticr3),
-    PH11: (ph11, 11, Input<Floating>, exticr3),
-    PH12: (ph12, 12, Input<Floating>, exticr4),
-    PH13: (ph13, 13, Input<Floating>, exticr4),
-    PH14: (ph14, 14, Input<Floating>, exticr4),
-    PH15: (ph15, 15, Input<Floating>, exticr4),
+    PH0: (ph0, 0, Input<Floating>),
+    PH1: (ph1, 1, Input<Floating>),
+    PH2: (ph2, 2, Input<Floating>),
+    PH3: (ph3, 3, Input<Floating>),
+    PH4: (ph4, 4, Input<Floating>),
+    PH5: (ph5, 5, Input<Floating>),
+    PH6: (ph6, 6, Input<Floating>),
+    PH7: (ph7, 7, Input<Floating>),
+    PH8: (ph8, 8, Input<Floating>),
+    PH9: (ph9, 9, Input<Floating>),
+    PH10: (ph10, 10, Input<Floating>),
+    PH11: (ph11, 11, Input<Floating>),
+    PH12: (ph12, 12, Input<Floating>),
+    PH13: (ph13, 13, Input<Floating>),
+    PH14: (ph14, 14, Input<Floating>),
+    PH15: (ph15, 15, Input<Floating>),
 ]);
 
 #[cfg(any(feature = "stm32f401"))]
 gpio!(GPIOH, gpioh, PH, 7, [
-    PH0: (ph0, 0, Input<Floating>, exticr1),
-    PH1: (ph1, 1, Input<Floating>, exticr1),
+    PH0: (ph0, 0, Input<Floating>),
+    PH1: (ph1, 1, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpioi")]
 gpio!(GPIOI, gpioi, PI, 8, [
-    PI0: (pi0, 0, Input<Floating>, exticr1),
-    PI1: (pi1, 1, Input<Floating>, exticr1),
-    PI2: (pi2, 2, Input<Floating>, exticr1),
-    PI3: (pi3, 3, Input<Floating>, exticr1),
-    PI4: (pi4, 4, Input<Floating>, exticr2),
-    PI5: (pi5, 5, Input<Floating>, exticr2),
-    PI6: (pi6, 6, Input<Floating>, exticr2),
-    PI7: (pi7, 7, Input<Floating>, exticr2),
-    PI8: (pi8, 8, Input<Floating>, exticr3),
-    PI9: (pi9, 9, Input<Floating>, exticr3),
-    PI10: (pi10, 10, Input<Floating>, exticr3),
-    PI11: (pi11, 11, Input<Floating>, exticr3),
-    PI12: (pi12, 12, Input<Floating>, exticr4),
-    PI13: (pi13, 13, Input<Floating>, exticr4),
-    PI14: (pi14, 14, Input<Floating>, exticr4),
-    PI15: (pi15, 15, Input<Floating>, exticr4),
+    PI0: (pi0, 0, Input<Floating>),
+    PI1: (pi1, 1, Input<Floating>),
+    PI2: (pi2, 2, Input<Floating>),
+    PI3: (pi3, 3, Input<Floating>),
+    PI4: (pi4, 4, Input<Floating>),
+    PI5: (pi5, 5, Input<Floating>),
+    PI6: (pi6, 6, Input<Floating>),
+    PI7: (pi7, 7, Input<Floating>),
+    PI8: (pi8, 8, Input<Floating>),
+    PI9: (pi9, 9, Input<Floating>),
+    PI10: (pi10, 10, Input<Floating>),
+    PI11: (pi11, 11, Input<Floating>),
+    PI12: (pi12, 12, Input<Floating>),
+    PI13: (pi13, 13, Input<Floating>),
+    PI14: (pi14, 14, Input<Floating>),
+    PI15: (pi15, 15, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpioj")]
 gpio!(GPIOJ, gpioj, PJ, 9, [
-    PJ0: (pj0, 0, Input<Floating>, exticr1),
-    PJ1: (pj1, 1, Input<Floating>, exticr1),
-    PJ2: (pj2, 2, Input<Floating>, exticr1),
-    PJ3: (pj3, 3, Input<Floating>, exticr1),
-    PJ4: (pj4, 4, Input<Floating>, exticr2),
-    PJ5: (pj5, 5, Input<Floating>, exticr2),
-    PJ6: (pj6, 6, Input<Floating>, exticr2),
-    PJ7: (pj7, 7, Input<Floating>, exticr2),
-    PJ8: (pj8, 8, Input<Floating>, exticr3),
-    PJ9: (pj9, 9, Input<Floating>, exticr3),
-    PJ10: (pj10, 10, Input<Floating>, exticr3),
-    PJ11: (pj11, 11, Input<Floating>, exticr3),
-    PJ12: (pj12, 12, Input<Floating>, exticr4),
-    PJ13: (pj13, 13, Input<Floating>, exticr4),
-    PJ14: (pj14, 14, Input<Floating>, exticr4),
-    PJ15: (pj15, 15, Input<Floating>, exticr4),
+    PJ0: (pj0, 0, Input<Floating>),
+    PJ1: (pj1, 1, Input<Floating>),
+    PJ2: (pj2, 2, Input<Floating>),
+    PJ3: (pj3, 3, Input<Floating>),
+    PJ4: (pj4, 4, Input<Floating>),
+    PJ5: (pj5, 5, Input<Floating>),
+    PJ6: (pj6, 6, Input<Floating>),
+    PJ7: (pj7, 7, Input<Floating>),
+    PJ8: (pj8, 8, Input<Floating>),
+    PJ9: (pj9, 9, Input<Floating>),
+    PJ10: (pj10, 10, Input<Floating>),
+    PJ11: (pj11, 11, Input<Floating>),
+    PJ12: (pj12, 12, Input<Floating>),
+    PJ13: (pj13, 13, Input<Floating>),
+    PJ14: (pj14, 14, Input<Floating>),
+    PJ15: (pj15, 15, Input<Floating>),
 ]);
 
 #[cfg(feature = "gpiok")]
 gpio!(GPIOK, gpiok, PK, 10, [
-    PK0: (pk0, 0, Input<Floating>, exticr1),
-    PK1: (pk1, 1, Input<Floating>, exticr1),
-    PK2: (pk2, 2, Input<Floating>, exticr1),
-    PK3: (pk3, 3, Input<Floating>, exticr1),
-    PK4: (pk4, 4, Input<Floating>, exticr2),
-    PK5: (pk5, 5, Input<Floating>, exticr2),
-    PK6: (pk6, 6, Input<Floating>, exticr2),
-    PK7: (pk7, 7, Input<Floating>, exticr2),
+    PK0: (pk0, 0, Input<Floating>),
+    PK1: (pk1, 1, Input<Floating>),
+    PK2: (pk2, 2, Input<Floating>),
+    PK3: (pk3, 3, Input<Floating>),
+    PK4: (pk4, 4, Input<Floating>),
+    PK5: (pk5, 5, Input<Floating>),
+    PK6: (pk6, 6, Input<Floating>),
+    PK7: (pk7, 7, Input<Floating>),
 ]);